backend/requirements.txt
Django==5.1.3
psycopg[binary]==3.2.9
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
django-cors-headers==4.4.0
Pillow==10.4.0
python-slugify==8.0.4
python-dotenv==1.0.1
drf-nested-routers==0.94.1
requests==2.32.4
